feat: changed repository parameters from Long to Integer

+++ src/main/java/hu/okrim/productratingapp/repository/ProductFlavourRepository.java
@@ -4,5 +4,5 @@ import hu.okrim.productratingapp.entity.Person;
 import hu.okrim.productratingapp.entity.ProductFlavour;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductFlavourRepository extends CrudRepository<ProductFlavour, Long> {
+public interface ProductFlavourRepository extends CrudRepository<ProductFlavour, Integer> {
 }
